Character and Race models complete
Completed the initial baselines for the CharacterModel and RaceModel. Will likely need to make adjustments as needed, but overall structure should be set.
Added method to DiceRollerFactory to provide average roll value for a given die.
Edited CharacterModel to use HitPointModel instead of just HitDiceModel.
Added test character to logging for crude testing purposes.
Minor bug fixes to Factories for typos.
Fixed bug with CanAddLanguage() and CanAddDefense() to account for null lists.