Refactored Attribute array into new Model
Refactored the List<AttributeModel> into a new Class AttributeArrayModel to better represent the fact that there is only ever the one array. This will simplify things further down the line and streamline adding creatures for the users.
Refactored SkillFactory with documentation for myself and to use the new AttributeArrayModel in place of the list.